refactor(bench): extract metrics and render modules from bench_serving
Next slices of the split:
- metrics.rs: dur_ms, percentiles, summarize_durations/counts,
  aggregate_tok_s, generated_token_hash/trace.
- render.rs: comfy_table table helpers and all render_* text renderers.

Both pull report types via `use crate::report::*`; main.rs re-exports each
module's pub(crate) items. Trimmed the now-unused comfy_table/io imports
from main.rs (Cell/CellAlignment stay — inline tables remain in the
snapshot/compare paths).

Pure code motion, no behavior change. Verified: default build clean (only
the pre-existing ParallelConfig kimi-gated unused-import warning remains).
