+++ openinfer-server/src/bin/bench_serving/main.rs
@@ -11,16 +11,13 @@
 
 use std::fmt::Write as _;
 use std::fs;
-use std::io::{IsTerminal, stdout};
 use std::path::{Path, PathBuf};
 use std::thread;
 use std::time::{Duration, Instant};
 
 use anyhow::{Context, Result, ensure};
 use clap::Parser;
-use comfy_table::modifiers::UTF8_ROUND_CORNERS;
-use comfy_table::presets::{ASCII_FULL_CONDENSED, UTF8_FULL_CONDENSED};
-use comfy_table::{Cell, CellAlignment, Table};
+use comfy_table::{Cell, CellAlignment};
 use cudarc::driver::Profiler;
 use cudarc::runtime::result::device as cuda_device;
 use log::{debug, info};
@@ -39,8 +36,12 @@ use tokio::sync::mpsc;
 use vllm_text::tokenizer::DynTokenizer;
 
 mod cli;
+mod metrics;
+mod render;
 mod report;
 use cli::*;
+use metrics::*;
+use render::*;
 use report::*;
 
 const SNAPSHOT_DIR: &str = concat!(env!("CARGO_MANIFEST_DIR"), "/../bench_snapshots");
@@ -64,359 +65,6 @@ const REGRESSION_TTFT_PCT: f64 = 3.0;
 const DEFAULT_REQUEST_PROMPT: &str = "Tell me a story";
 const DEFAULT_CURVE_PROMPT_LEN: usize = 512;
 const SYNTHETIC_PATTERN: &str = "token_id = 100 + (idx % 1000)";
-
-fn dur_ms(d: Duration) -> f64 {
-    d.as_secs_f64() * 1000.0
-}
-
-fn percentiles(sorted: &[Duration]) -> (Duration, Duration, Duration, Duration, Duration) {
-    assert!(!sorted.is_empty());
-    let n = sorted.len();
-    let sum: Duration = sorted.iter().sum();
-    let avg = sum / n as u32;
-    let p = |pct: f64| sorted[((pct / 100.0) * (n - 1) as f64).round() as usize];
-    (avg, p(50.0), p(95.0), p(99.0), sorted[n - 1])
-}
-
-fn summarize_durations(samples: &[Duration]) -> DurationStats {
-    let mut sorted = samples.to_vec();
-    sorted.sort();
-    let (avg, p50, p95, p99, max) = percentiles(&sorted);
-    DurationStats {
-        avg_ms: dur_ms(avg),
-        p50_ms: dur_ms(p50),
-        p95_ms: dur_ms(p95),
-        p99_ms: dur_ms(p99),
-        max_ms: dur_ms(max),
-        samples: sorted.len(),
-    }
-}
-
-fn summarize_counts(samples: &[usize]) -> CountStats {
-    assert!(!samples.is_empty());
-    let min = *samples.iter().min().unwrap();
-    let max = *samples.iter().max().unwrap();
-    let sum: usize = samples.iter().sum();
-    CountStats {
-        min,
-        max,
-        avg: sum as f64 / samples.len() as f64,
-        samples: samples.len(),
-    }
-}
-
-fn aggregate_tok_s(tokens: usize, total: Duration) -> Option<f64> {
-    if tokens == 0 || total.is_zero() {
-        None
-    } else {
-        Some(tokens as f64 / total.as_secs_f64())
-    }
-}
-
-fn generated_token_hash(tokens: &[u32]) -> String {
-    let mut hash = 0xcbf2_9ce4_8422_2325_u64;
-    for token in tokens {
-        for byte in token.to_le_bytes() {
-            hash ^= u64::from(byte);
-            hash = hash.wrapping_mul(0x0100_0000_01b3);
-        }
-    }
-    format!("{hash:016x}")
-}
-
-fn generated_token_trace(tokens: &[u32]) -> GeneratedTokenTrace {
-    GeneratedTokenTrace {
-        hash: generated_token_hash(tokens),
-        prefix: tokens.iter().copied().take(16).collect(),
-        len: tokens.len(),
-    }
-}
-
-fn new_table() -> Table {
-    let mut table = Table::new();
-    if stdout().is_terminal() {
-        table.load_preset(UTF8_FULL_CONDENSED);
-        table.apply_modifier(UTF8_ROUND_CORNERS);
-    } else {
-        table.load_preset(ASCII_FULL_CONDENSED);
-    }
-    table
-}
-
-fn key_cell(label: impl Into<String>) -> Cell {
-    Cell::new(label.into())
-}
-
-fn value_cell(value: impl Into<String>) -> Cell {
-    Cell::new(value.into())
-}
-
-fn numeric_cell(value: impl Into<String>) -> Cell {
-    Cell::new(value.into()).set_alignment(CellAlignment::Right)
-}
-
-fn format_rate(value: Option<f64>) -> String {
-    value.map_or_else(|| "-".to_string(), |v| format!("{v:.2}"))
-}
-
-fn format_duration_ms(value: f64) -> String {
-    format!("{value:.2}")
-}
-
-fn format_count_avg(value: f64) -> String {
-    format!("{value:.2}")
-}
-
-fn push_table(out: &mut String, table: &Table) {
-    out.push_str(&table.to_string());
-    out.push('\n');
-}
-
-fn render_run_summary(report: &RunInfo) -> Table {
-    let mut table = new_table();
-    table.add_row(vec![
-        key_cell("model"),
-        value_cell(format!("{} ({})", report.model_path, report.model_type)),
-    ]);
-    table.add_row(vec![
-        key_cell("cuda_graph"),
-        value_cell(report.cuda_graph.to_string()),
-    ]);
-    table.add_row(vec![
-        key_cell("load_ms"),
-        numeric_cell(format_duration_ms(report.load_ms)),
-    ]);
-    if let Some(label) = &report.label {
-        table.add_row(vec![key_cell("label"), value_cell(label.clone())]);
-    }
-    table
-}
-
-fn render_request_meta(report: &RequestReport) -> Table {
-    let mut table = render_run_summary(&report.run);
-    table.add_row(vec![
-        key_cell("prompt_source"),
-        value_cell(report.workload.prompt.source.clone()),
-    ]);
-    table.add_row(vec![
-        key_cell("prompt_tokens"),
-        numeric_cell(report.workload.prompt.prompt_tokens.to_string()),
-    ]);
-    if let Some(preview) = &report.workload.prompt.prompt_preview {
-        table.add_row(vec![
-            key_cell("prompt"),
-            value_cell(format!("\"{preview}\"")),
-        ]);
-    }
-    table.add_row(vec![
-        key_cell("output_len"),
-        numeric_cell(report.workload.output_len.to_string()),
-    ]);
-    table.add_row(vec![
-        key_cell("warmup / iters"),
-        value_cell(format!(
-            "{} / {}",
-            report.workload.warmup, report.workload.iters
-        )),
-    ]);
-    table.add_row(vec![
-        key_cell("seed"),
-        numeric_cell(report.workload.seed.to_string()),
-    ]);
-    table
-}
-
-fn render_duration_table(rows: Vec<(String, DurationStats)>) -> Table {
-    let mut table = new_table();
-    table.set_header(vec![
-        Cell::new("metric"),
-        Cell::new("avg_ms").set_alignment(CellAlignment::Right),
-        Cell::new("p50_ms").set_alignment(CellAlignment::Right),
-        Cell::new("p95_ms").set_alignment(CellAlignment::Right),
-        Cell::new("p99_ms").set_alignment(CellAlignment::Right),
-        Cell::new("max_ms").set_alignment(CellAlignment::Right),
-        Cell::new("samples").set_alignment(CellAlignment::Right),
-    ]);
-    for (label, stats) in rows {
-        table.add_row(vec![
-            key_cell(label),
-            numeric_cell(format_duration_ms(stats.avg_ms)),
-            numeric_cell(format_duration_ms(stats.p50_ms)),
-            numeric_cell(format_duration_ms(stats.p95_ms)),
-            numeric_cell(format_duration_ms(stats.p99_ms)),
-            numeric_cell(format_duration_ms(stats.max_ms)),
-            numeric_cell(stats.samples.to_string()),
-        ]);
-    }
-    table
-}
-
-fn render_request_summary(report: &RequestReport) -> Table {
-    let mut table = new_table();
-    table.set_header(vec![
-        Cell::new("metric"),
-        Cell::new("value").set_alignment(CellAlignment::Right),
-    ]);
-    table.add_row(vec![
-        key_cell("generated_tokens_avg"),
-        numeric_cell(format_count_avg(report.metrics.generated_tokens.avg)),
-    ]);
-    table.add_row(vec![
-        key_cell("generated_tokens_min"),
-        numeric_cell(report.metrics.generated_tokens.min.to_string()),
-    ]);
-    table.add_row(vec![
-        key_cell("generated_tokens_max"),
-        numeric_cell(report.metrics.generated_tokens.max.to_string()),
-    ]);
-    table.add_row(vec![
-        key_cell("generated_token_runs"),
-        numeric_cell(report.metrics.generated_tokens.samples.to_string()),
-    ]);
-    table.add_row(vec![
-        key_cell("request_tok_s"),
-        numeric_cell(format_rate(report.metrics.request_tok_s)),
-    ]);
-    table.add_row(vec![
-        key_cell("decode_tok_s"),
-        numeric_cell(format_rate(report.metrics.decode_tok_s)),
-    ]);
-    table
-}
-
-fn render_matrix_meta(report: &MatrixReport) -> Table {
-    let mut table = render_run_summary(&report.run);
-    table.add_row(vec![
-        key_cell("prompt_lens"),
-        value_cell(
-            report
-                .workload
-                .prompt_lens
-                .iter()
-                .map(std::string::ToString::to_string)
-                .collect::<Vec<_>>()
-                .join(","),
-        ),
-    ]);
-    table.add_row(vec![
-        key_cell("output_lens"),
-        value_cell(
-            report
-                .workload
-                .output_lens
-                .iter()
-                .map(std::string::ToString::to_string)
-                .collect::<Vec<_>>()
-                .join(","),
-        ),
-    ]);
-    table.add_row(vec![
-        key_cell("synthetic_pattern"),
-        value_cell(report.workload.synthetic_pattern),
-    ]);
-    table.add_row(vec![
-        key_cell("warmup / iters"),
-        value_cell(format!(
-            "{} / {}",
-            report.workload.warmup, report.workload.iters
-        )),
-    ]);
-    table.add_row(vec![
-        key_cell("seed"),
-        numeric_cell(report.workload.seed.to_string()),
-    ]);
-    table
-}
-
-fn render_matrix_table(report: &MatrixReport) -> Table {
-    let mut table = new_table();
-    table.set_header(vec![
-        Cell::new("prompt_tok").set_alignment(CellAlignment::Right),
-        Cell::new("output_tok").set_alignment(CellAlignment::Right),
-        Cell::new("ttft_avg").set_alignment(CellAlignment::Right),
-        Cell::new("ttft_p95").set_alignment(CellAlignment::Right),
-        Cell::new("e2e_avg").set_alignment(CellAlignment::Right),
-        Cell::new("req_tok/s").set_alignment(CellAlignment::Right),
-        Cell::new("decode_tok/s").set_alignment(CellAlignment::Right),
-        Cell::new("gen_avg").set_alignment(CellAlignment::Right),
-    ]);
-    for cell in &report.cells {
-        table.add_row(vec![
-            numeric_cell(cell.prompt_len.to_string()),
-            numeric_cell(cell.output_len.to_string()),
-            numeric_cell(format_duration_ms(cell.ttft_ms.avg_ms)),
-            numeric_cell(format_duration_ms(cell.ttft_ms.p95_ms)),
-            numeric_cell(format_duration_ms(cell.e2e_ms.avg_ms)),
-            numeric_cell(format_rate(cell.request_tok_s)),
-            numeric_cell(format_rate(cell.decode_tok_s)),
-            numeric_cell(format_count_avg(cell.generated_tokens.avg)),
-        ]);
-    }
-    table
-}
-
-fn render_curve_meta(report: &CurveReport) -> Table {
-    let mut table = render_run_summary(&report.run);
-    table.add_row(vec![
-        key_cell("prompt_source"),
-        value_cell(report.workload.prompt.source.clone()),
-    ]);
-    table.add_row(vec![
-        key_cell("prompt_tokens"),
-        numeric_cell(report.workload.prompt.prompt_tokens.to_string()),
-    ]);
-    if let Some(preview) = &report.workload.prompt.prompt_preview {
-        table.add_row(vec![
-            key_cell("prompt"),
-            value_cell(format!("\"{preview}\"")),
-        ]);
-    }
-    table.add_row(vec![
-        key_cell("output_len"),
-        numeric_cell(report.workload.output_len.to_string()),
-    ]);
-    table.add_row(vec![
-        key_cell("window"),
-        numeric_cell(report.workload.window.to_string()),
-    ]);
-    table.add_row(vec![
-        key_cell("warmup / iters"),
-        value_cell(format!(
-            "{} / {}",
-            report.workload.warmup, report.workload.iters
-        )),
-    ]);
-    table.add_row(vec![
-        key_cell("seed"),
-        numeric_cell(report.workload.seed.to_string()),
-    ]);
-    table
-}
-
-fn render_curve_table(report: &CurveReport) -> Table {
-    let mut table = new_table();
-    table.set_header(vec![
-        Cell::new("ctx_range"),
-        Cell::new("avg_ms").set_alignment(CellAlignment::Right),
-        Cell::new("p50_ms").set_alignment(CellAlignment::Right),
-        Cell::new("p95_ms").set_alignment(CellAlignment::Right),
-        Cell::new("p99_ms").set_alignment(CellAlignment::Right),
-        Cell::new("tok/s").set_alignment(CellAlignment::Right),
-        Cell::new("samples").set_alignment(CellAlignment::Right),
-    ]);
-    for window in &report.windows {
-        table.add_row(vec![
-            value_cell(format!("{}-{}", window.ctx_start, window.ctx_end)),
-            numeric_cell(format_duration_ms(window.tpot_ms.avg_ms)),
-            numeric_cell(format_duration_ms(window.tpot_ms.p50_ms)),
-            numeric_cell(format_duration_ms(window.tpot_ms.p95_ms)),
-            numeric_cell(format_duration_ms(window.tpot_ms.p99_ms)),
-            numeric_cell(format_rate(window.decode_tok_s)),
-            numeric_cell(window.tpot_ms.samples.to_string()),
-        ]);
-    }
-    table
-}
 
 fn truncate_preview(text: &str, limit: usize) -> String {
     let one_line = text.replace('\n', "\\n");
